backend/requirements.txt
# Core
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI / Email Parsing
anthropic>=0.18.0
google-genai>=0.4.0
httpx>=0.26.0

# Email
resend>=0.8.0
aiosmtplib>=3.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.114.0

# Utilities
python-dotenv>=1.0.0
bleach>=6.1.0

# Monitoring
sentry-sdk[fastapi]>=1.40.0

# Background Jobs
apscheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
